move group and modal component to components instead of containers
