+++ app/javascript/react_app/components/group.jsx
