+++ app/javascript/react_app/components/modal.jsx
