+++ app/javascript/react_app/containers/group_list.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { fetchGroups } from '../actions';
-import Group from './group';
+import Group from '../components/group';
 
 class GroupList extends Component {
   componentDidMount() {
